backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
psycopg2-binary==2.9.12
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
ultralytics
opencv-python
facenet-pytorch
umpy
torch
torchvision

pandas
openpyxl
fpdf2
